Moved configuration to libconfig
libconfig will be responsible for handling all the configuration state
of bear
