+++ source/CMakeLists.txt
@@ -80,6 +80,7 @@ add_subdirectory(libflags)
 add_subdirectory(libshell)
 add_subdirectory(libsys)
 add_subdirectory(libmain)
+add_subdirectory(libconfig)
 add_subdirectory(intercept)
 add_subdirectory(citnames)
 add_subdirectory(bear)
