+++ source/libconfig/include/libconfig/Configuration.h
